refactor event dropdown component
