+++ pages/events/index.vue
@@ -12,27 +12,20 @@
       </div>
       <div class="event-page__filters-region">
         <client-only>
-          <!-- <cv-multi-select
-            :theme="theme"
-            :label="labelRegions"
-            :options="regionOptions"
-            :value="getCheckedFilters('regionFilters')"
-            :selection-feedback="feedback"
-            @change="selectedOptions => updateWholeFilter('regionFilters', selectedOptions)"
-          /> -->
-          <EventDropdown />
+          <EventMultiSelect
+            :label="regionsLabel"
+            :options="regionsOptions"
+            :filter-type="regionsFilters"
+          />
         </client-only>
       </div>
       <div class="event-page__filters-type">
         <client-only>
-          <!-- <cv-multi-select
-            :theme="theme"
-            :label="labelTypes"
-            :options="typeOptions"
-            :value="getCheckedFilters('typeFilters')"
-            :selection-feedback="feedback"
-            @change="selectedOptions => updateWholeFilter('typeFilters', selectedOptions)"
-          /> -->
+          <EventMultiSelect
+            :label="typesLabel"
+            :options="typesOptions"
+            :filter-type="typesFilters"
+          />
         </client-only>
       </div>
       <div class="event-page__event-index">
@@ -124,8 +117,7 @@ import AppCard from '~/components/ui/AppCard.vue'
 import TheEventsHeader from '~/components/events/TheEventsHeader.vue'
 import AppCta from '~/components/ui/AppCta.vue'
 import LandingCta from '~/components/landing/LandingCta.vue'
-import EventDropdown from '~/components/events/EventDropdown.vue'
-
+import EventMultiSelect from '~/components/events/EventMultiSelect.vue'
 import {
   CommunityEvent,
   WORLD_REGION_OPTIONS,
@@ -141,7 +133,7 @@ import { EVENT_REQUEST_LINK } from '~/constants/appLinks'
     LandingCta,
     AppCard,
     TheEventsHeader,
-    EventDropdown
+    EventMultiSelect
   },
 
   head () {
@@ -186,12 +178,12 @@ export default class extends QiskitPage {
   }
 
   // multiselect
-  regionOptions = this.getOptions(this.regions)
-  typeOptions = this.getOptions(this.types)
-  theme: string = 'light'
-  labelRegions: string = 'Locations'
-  labelTypes: string = 'Types'
-  feedback: string = 'fixed'
+  regionsOptions = this.getOptions(this.regions)
+  typesOptions = this.getOptions(this.types)
+  regionsLabel: string = 'Locations'
+  typesLabel: string = 'Types'
+  regionsFilters: string = 'regionFilters'
+  typesFilters: string = 'typeFilters'
 
   get noEvents (): boolean {
     return (this as any).filteredEvents.length === 0
@@ -199,17 +191,6 @@ export default class extends QiskitPage {
 
   getOptions (optionsList): Array<EventMultiSelectOption> {
     return optionsList.map(item => ({ label: item, value: item, name: item }))
-  }
-
-  getCheckedFilters (filter) {
-    return (this as any)[filter]
-  }
-
-  updateWholeFilter (filter: string, filterValues: string[]): void {
-    const { commit } = this.$store
-    const payload = { filter, filterValues }
-
-    commit('updateFilterSet', payload)
   }
 
   isFilterChecked (filter: string, filterValue: string): Array<CommunityEvent> {
@@ -330,71 +311,6 @@ export default class extends QiskitPage {
       display: none;
     }
 
-  }
-
-  &__filters-region,
-  &__filters-type {
-    @include mq($from: medium) {
-      display: none;
-    }
-
-    .bx--list-box--light,
-    .bx--list-box__menu {
-      background-color: $white;
-      border-bottom-color: $gray-20;
-    }
-
-    .bx--list-box__label,
-    .bx--list-box__menu-item,
-    .bx--list-box__menu-item--highlighted .bx--list-box__menu-item__option,
-    .bx--list-box__menu-item__option {
-      color: $white-text-01;
-    }
-
-    .bx--list-box__menu-item {
-      background-color: $cool-gray-10;
-      color: $white-text-01;
-    }
-
-    .bx--list-box__menu-icon > svg {
-      fill: $black-100;
-    }
-
-    .bx--list-box--expanded:hover.bx--list-box--light:hover {
-      background-color: $cool-gray-10;
-    }
-
-    .bx--checkbox-label::before {
-      border: 1px solid $black-100;
-    }
-
-    .bx--list-box__menu-item--highlighted {
-      background-color: $cool-gray-20;
-    }
-
-    .bx--tag--filter {
-      background-color: $purple-70;
-      color: $white;
-    }
-
-    .bx--tag--high-contrast .bx--tag__close-icon:hover {
-      background-color: $purple-70;
-    }
-
-    .bx--checkbox:checked + .bx--checkbox-label::before {
-      background-color: $black-100;
-      border-color: $black-100;
-      border-width: 1px;
-    }
-
-    .bx--checkbox:checked + .bx--checkbox-label::after {
-      border-left: 2px solid $white;
-      border-bottom: 2px solid $white;
-    }
-
-    .bx--list-box__menu-item:hover .bx--list-box__menu-item__option {
-      color: $black-100;
-    }
   }
 
   &__filters-type {
